Fix immutability bug

+++ Gordon360/Controllers/RecIM/SportsController.cs
@@ -66,6 +66,7 @@ namespace Gordon360.Controllers.RecIM
         [StateYourBusiness(operation = Operation.DELETE, resource = Resource.RECIM_SPORT)]
         public async Task<ActionResult<SportViewModel>> DeleteSport(int sportID)
         {
+            if (sportID == 0) return UnprocessableEntity("Default sport cannot be modified");
             var res = await _sportService.DeleteSportAsync(sportID);
             return Ok(res);
         }
